+++ lightwood/api/high_level.py
@@ -1,5 +1,4 @@
 import os
-from types import ModuleType
 from typing import Union
 import dill
 import pandas as pd
@@ -8,12 +7,8 @@ from dataprep_ml.insights import statistical_analysis
 from type_infer.infer import infer_types
 from lightwood.api.predictor import PredictorInterface
 from lightwood.api.json_ai import generate_json_ai
-import tempfile
-from lightwood.helpers.codegen import code_from_json_ai as _code_from_json_ai
-import importlib.util
+from lightwood.helpers.codegen import code_from_json_ai as _code_from_json_ai, _module_from_code, _predictor_from_code
 import sys
-import random
-import string
 import gc
 import time
 from lightwood.helpers.log import log
@@ -107,10 +102,7 @@ def predictor_from_code(code: str) -> PredictorInterface:
 
     :returns: A lightwood ``Predictor`` object
     """
-    module_name = ''.join(random.choices(string.ascii_uppercase + string.digits, k=12))
-    module_name += str(time.time()).replace('.', '')
-    predictor = _module_from_code(code, module_name).Predictor()
-    return predictor
+    return _predictor_from_code(code)
 
 
 def code_from_problem(df: pd.DataFrame, problem_definition: Union[ProblemDefinition, dict]) -> str:
@@ -160,31 +152,6 @@ def predictor_from_state(state_file: str, code: str = None) -> PredictorInterfac
             predictor = dill.load(fp)
 
     return predictor
-
-
-def _module_from_code(code: str, module_name: str) -> ModuleType:
-    """
-    Create a python module (containing the generated ``Predictor`` class) from the code. This is both a python object and an associated temporary file on your filesystem
-
-    :param code: The ``Predictor``'s code in text form
-    :param module_name: The name of the newly created module
-
-    :returns: A python module object
-    """ # noqa
-    dirname = tempfile.gettempdir()
-    filename = os.urandom(24).hex() + str(time.time()).replace('.', '') + '.py'
-    path = os.path.join(dirname, filename)
-    if 'LIGHTWOOD_DEV_SAVE_TO' in os.environ:
-        path = os.environ['LIGHTWOOD_DEV_SAVE_TO']
-
-    with open(path, 'wb') as fp:
-        fp.write(code.encode('utf-8'))
-        spec = importlib.util.spec_from_file_location(module_name, fp.name)
-        temp_module = importlib.util.module_from_spec(spec)
-        sys.modules[module_name] = temp_module
-        spec.loader.exec_module(temp_module)
-
-    return temp_module
 
 
 def predictor_from_json_ai(json_ai: JsonAI) -> PredictorInterface:
